fixes bug where a var change doesn't trigger a cache update
