+++ src/datasource.ts
@@ -2,7 +2,7 @@ import { DataSourceInstanceSettings, CoreApp, ScopedVars, DataQueryResponse, Dat
 import { DataSourceWithBackend, getTemplateSrv } from '@grafana/runtime';
 import { SiftQuery, SiftDataSourceOptions, DEFAULT_QUERY, AssetQuery, AssetGrafanaVariable } from './types';
 import { SiftVariableSupport } from 'variables';
-import { filterQueryBeforeRequest } from './utils';
+import { filterQueryBeforeRequest, replaceTemplateVariablesInQuery } from './utils';
 import { Observable, from } from 'rxjs';
 import { switchMap } from 'rxjs/operators';
 import { SiftDataSourceCache } from './datasourceCache';
@@ -51,77 +51,6 @@ export class SiftDataSource extends DataSourceWithBackend<SiftQuery, SiftDataSou
   applyTemplateVariables(query: SiftQuery, scopedVars: ScopedVars): SiftQuery {
     // First filter any queries that aren't fully defined
     const filteredQuery = filterQueryBeforeRequest(query);
-
-    const templateSrv = getTemplateSrv();
-
-    function getValuesForVariable(name: string): string[] {
-      const values: string[] = [];
-      templateSrv.replace(name, {}, (value: string | string[]) => {
-        if (Array.isArray(value)) {
-          values.push(...value);
-        } else {
-          values.push(value);
-        }
-      });
-      return values;
-    }
-
-    return {
-      ...filteredQuery,
-      channelDataQueries: filteredQuery.channelDataQueries?.map((cdq) => {
-        return {
-          ...cdq,
-          assetQueries: cdq.assetQueries?.reduce((acc: AssetQuery[], aq: AssetQuery) => {
-            // If we have a dashboard variable, handle it separately
-            if (aq.dashboardVariableName) {
-              const dashboardVarValues = getValuesForVariable(aq.dashboardVariableName);
-              if (dashboardVarValues) {
-                dashboardVarValues.forEach((v) => {
-                  acc.push({
-                    assetId: v,
-                    dashboardVariableName: aq.dashboardVariableName,
-                  });
-                });
-              }
-            } else {
-              acc.push({
-                ...aq,
-                assetId: templateSrv.replace(aq.assetId || ''),
-                assetName: templateSrv.replace(aq.assetName || ''),
-              });
-            }
-            return acc;
-          }, []),
-          runQueries: cdq.runQueries?.map((rq) => {
-            return {
-              ...rq,
-              runId: templateSrv.replace(rq.runId || ''),
-              runName: templateSrv.replace(rq.runName || ''),
-            };
-          }),
-          channelQueries: cdq.channelQueries?.map((cq) => {
-            return {
-              ...cq,
-              channelId: templateSrv.replace(cq.channelId || ''),
-              channelName: templateSrv.replace(cq.channelName || ''),
-            };
-          }),
-          calculatedChannelQueries: cdq.calculatedChannelQueries?.map((cc) => {
-            return {
-              ...cc,
-              name: templateSrv.replace(cc.name || ''),
-              expression: templateSrv.replace(cc.expression || ''),
-              channelReferences: cc.channelReferences?.map((cr) => {
-                return {
-                  ...cr,
-                  channelId: templateSrv.replace(cr.channelId || ''),
-                  channelName: templateSrv.replace(cr.channelName || ''),
-                };
-              }),
-            };
-          }),
-        };
-      }),
-    };
+    return replaceTemplateVariablesInQuery(filteredQuery);
   }
 }
